feat(estructura frontend): modulariza el front end en componentes para que sea mas mantenible
