+++ src/frontend/src/components/Loader.tsx
@@ -1,0 +1,12 @@
+// src/components/ui/Loader.tsx
+import React from 'react';
+
+const Loader: React.FC = () => {
+  return (
+    <div className="flex justify-center items-center h-screen">
+      <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-blue-600"></div>
+    </div>
+  );
+};
+
+export default Loader;
