+++ src/frontend/src/hooks/useFetchData.ts
@@ -1,0 +1,26 @@
+import { useState, useEffect } from 'react';
+import { DataResponse } from '../types';
+
+export const useFetchData = (url: string) => {
+  const [data, setData] = useState<DataResponse | null>(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<Error | null>(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await fetch(url);
+        const jsonData: DataResponse = await response.json();
+        setData(jsonData);
+        setLoading(false);
+      } catch (err) {
+        setError(err instanceof Error ? err : new Error('An unknown error occurred'));
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, [url]);
+
+  return { data, loading, error };
+};
