+++ src/frontend/src/types/index.ts
@@ -1,0 +1,11 @@
+export interface DataResponse {
+    message: string;
+  }
+  
+  export interface FeaturedItem {
+    title: string;
+    provider: string;
+    category: string;
+    date: string;
+    image: string;
+  }
