+++ src/frontend/src/utils/constants.ts
@@ -1,0 +1,12 @@
+import { FeaturedItem } from "../types";
+
+export const FEATURED_ITEMS: FeaturedItem[] = [
+    {
+      title: "Excel - de básico a Intermedio",
+      provider: "Impulsa Academy",
+      category: "Herramientas",
+      date: "Abierto hasta Mar 31, 25",
+      image: "/excel-course.jpg"
+    },
+    // ... other items from the original code
+  ];
